Refactor, cleanup, fixes and tweaks
