+++ src/main/kotlin/io/github/inductiveautomation/kindling/thread/comparison/ThreadComparisonPane.kt
@@ -1,49 +1,29 @@
-package io.github.inductiveautomation.kindling.thread
+package io.github.inductiveautomation.kindling.thread.comparison
 
 import com.formdev.flatlaf.extras.FlatSVGIcon
-import com.formdev.flatlaf.extras.components.FlatButton
-import com.formdev.flatlaf.extras.components.FlatLabel
 import com.jidesoft.swing.StyledLabel
-import io.github.inductiveautomation.kindling.core.DetailsPane
-import io.github.inductiveautomation.kindling.core.Kindling.Preferences.Advanced.HyperlinkStrategy
-import io.github.inductiveautomation.kindling.core.Kindling.Preferences.General.UseHyperlinks
-import io.github.inductiveautomation.kindling.thread.MultiThreadView.Companion.toDetail
-import io.github.inductiveautomation.kindling.thread.MultiThreadViewer.ShowEmptyValues
-import io.github.inductiveautomation.kindling.thread.MultiThreadViewer.ShowNullThreads
+import io.github.inductiveautomation.kindling.core.Kindling
+import io.github.inductiveautomation.kindling.thread.MultiThreadViewer
 import io.github.inductiveautomation.kindling.thread.model.Thread
 import io.github.inductiveautomation.kindling.thread.model.ThreadLifespan
 import io.github.inductiveautomation.kindling.utils.Action
-import io.github.inductiveautomation.kindling.utils.diff.DiffView
-import io.github.inductiveautomation.kindling.utils.EDT_SCOPE
-import io.github.inductiveautomation.kindling.utils.FlatScrollPane
 import io.github.inductiveautomation.kindling.utils.add
 import io.github.inductiveautomation.kindling.utils.attachPopupMenu
 import io.github.inductiveautomation.kindling.utils.dismissOnEscape
-import io.github.inductiveautomation.kindling.utils.escapeHtml
 import io.github.inductiveautomation.kindling.utils.getAll
 import io.github.inductiveautomation.kindling.utils.jFrame
-import io.github.inductiveautomation.kindling.utils.scrollToTop
 import io.github.inductiveautomation.kindling.utils.style
-import io.github.inductiveautomation.kindling.utils.tag
-import io.github.inductiveautomation.kindling.utils.toBodyLine
-import net.miginfocom.swing.MigLayout
 import java.awt.Color
-import java.awt.Dimension
 import java.awt.Font
-import java.text.DecimalFormat
 import java.util.EventListener
 import javax.swing.JButton
-import javax.swing.JCheckBox
 import javax.swing.JLabel
-import javax.swing.JOptionPane
 import javax.swing.JPanel
 import javax.swing.JPopupMenu
-import javax.swing.JTextPane
 import javax.swing.UIManager
 import javax.swing.event.EventListenerList
-import javax.swing.event.HyperlinkEvent
 import kotlin.properties.Delegates
-import kotlinx.coroutines.launch
+import net.miginfocom.swing.MigLayout
 
 class ThreadComparisonPane(
     totalThreadDumps: Int,
@@ -89,8 +69,6 @@ class ThreadComparisonPane(
         }
     }
 
-    private val header = HeaderPanel()
-
     private var selectionModel = ThreadSelectionModel()
 
     init {
@@ -104,19 +82,19 @@ class ThreadComparisonPane(
             }
         }
 
-        ShowNullThreads.addChangeListener {
+        MultiThreadViewer.ShowNullThreads.addChangeListener {
             for (container in threadContainers) {
                 container.updateThreadInfo()
             }
 
             footer?.reset()
         }
-        ShowEmptyValues.addChangeListener {
+        MultiThreadViewer.ShowEmptyValues.addChangeListener {
             for (container in threadContainers) {
                 container.updateThreadInfo()
             }
         }
-        UseHyperlinks.addChangeListener {
+        Kindling.Preferences.General.UseHyperlinks.addChangeListener {
             for (container in threadContainers) {
                 container.updateThreadInfo()
             }
@@ -252,7 +230,7 @@ class ThreadComparisonPane(
         }
     }
 
-    inner class FooterPanel : JPanel(MigLayout("fill, ins 3")) {
+    private inner class FooterPanel : JPanel(MigLayout("fill, ins 3")) {
         private val nextButton = JButton(nextIcon).apply {
             addActionListener { selectNext() }
             isEnabled = false
@@ -282,6 +260,7 @@ class ThreadComparisonPane(
 
                 recalculateConstraints()
             }
+
         private val canSelectNext: Boolean
             get() {
                 if (selectedIndices.isEmpty()) return false
@@ -355,264 +334,11 @@ class ThreadComparisonPane(
         fun asList(): List<ThreadContainer> = internal.sortedBy { it.index }
     }
 
-    private class BlockerButton : FlatButton() {
-        var blocker: Int? = null
-            set(value) {
-                isVisible = value != null
-                text = value?.toString()
-                field = value
-            }
-
-        init {
-            icon = blockedIcon
-            toolTipText = "Jump to blocking thread"
-            isVisible = false
-        }
-    }
-
-    private class DetailContainer(
-        val prefix: String,
-        collapsed: Boolean = true,
-    ) : JPanel(MigLayout("ins 0, fill, flowy, hidemode 3")) {
-        var isCollapsed: Boolean = collapsed
-            set(value) {
-                field = value
-                scrollPane.isVisible = !value
-                headerButton.icon = if (value) expandIcon else collapseIcon
-
-                // Preferred size needs to be recalculated or something like that
-                EDT_SCOPE.launch {
-                    revalidate()
-                    repaint()
-                }
-            }
-
-        var itemCount = 0
-            set(value) {
-                headerButton.text = "$prefix: $value"
-                field = value
-            }
-
-        var isHightlighted: Boolean = false
-            set(value) {
-                field = value
-                if (value) {
-                    headerButton.foreground = threadHighlightColor
-                } else {
-                    headerButton.foreground = UIManager.getColor("Button.foreground")
-                }
-            }
-
-        private val headerButton = object : JButton(prefix, if (collapsed) expandIcon else collapseIcon) {
-            init {
-                horizontalTextPosition = LEFT
-                horizontalAlignment = LEFT
-
-                addActionListener { isCollapsed = !isCollapsed }
-            }
-
-            // This effectively right-aligns the icon
-            override fun getIconTextGap(): Int {
-                val fm = getFontMetrics(font)
-                return size.width - insets.left - insets.right - icon.iconWidth - fm.stringWidth(text)
-            }
-        }
-
-        private val textArea = JTextPane().apply {
-            isEditable = false
-            contentType = "text/html"
-
-            addHyperlinkListener { event ->
-                if (event.eventType == HyperlinkEvent.EventType.ACTIVATED) {
-                    HyperlinkStrategy.currentValue.handleEvent(event)
-                }
-            }
-        }
-
-        private val scrollPane = FlatScrollPane(textArea) {
-            horizontalScrollBar.preferredSize = Dimension(0, 10)
-            verticalScrollBar.preferredSize = Dimension(10, 0)
-
-            isVisible = !collapsed
-        }
-
-        var text: String?
-            get() = textArea.text
-            set(value) {
-                textArea.text = value
-            }
-
-        init {
-            add(headerButton, "growx, top, wmax 100%")
-            add(scrollPane, "push, grow, top, wmax 100%")
-        }
-
-        companion object {
-            private val collapseIcon = FlatSVGIcon("icons/bx-chevrons-up.svg")
-            private val expandIcon = FlatSVGIcon("icons/bx-chevrons-down.svg")
-        }
-    }
-
-    private class ThreadContainer(
-        private val version: String
-    ) : JPanel(MigLayout("fill, flowy, hidemode 3, gapy 5, ins 0")) {
-        var thread: Thread? by Delegates.observable(null) { _, _, _ ->
-            updateThreadInfo()
-        }
-
-        var highlightCpu: Boolean = false
-        var highlightStacktrace: Boolean = true
-
-        val isViewable: Boolean
-            get() = thread != null || ShowNullThreads.currentValue
-
-        var isSelected: Boolean = true
-            set(value) {
-                field = value
-                isVisible = isViewable && value
-            }
-
-        private val titleLabel = FlatLabel()
-        private val detailsButton = FlatButton().apply {
-            icon = detailIcon
-            toolTipText = "Open in details popup"
-            addActionListener {
-                thread?.let {
-                    jFrame("Thread ${it.name} Details", 900, 500) {
-                        contentPane = DetailsPane(listOf(it.toDetail(version)))
-                        dismissOnEscape()
-                    }
-                }
-            }
-        }
-
-        val blockerButton = BlockerButton()
-
-        val diffCheckBox = JCheckBox("Diff").apply {
-            isVisible = false
-            addActionListener {
-                this@ThreadContainer.firePropertyChange("selected", !isSelected, isSelected)
-            }
-        }
-
-        private val monitors = DetailContainer("Locked Monitors")
-        private val synchronizers = DetailContainer("Synchronizers")
-        private val stacktrace = DetailContainer("Stacktrace", false)
-
-        init {
-            add(
-                JPanel(MigLayout("fill, ins 5, hidemode 3")).apply {
-                    add(detailsButton)
-                    add(titleLabel, "push, grow, gapleft 8")
-                    add(diffCheckBox)
-                    add(blockerButton)
-                }, "wmax 100%"
-            )
-
-            // Ensure that the top two don't get pushed below their preferred size.
-            add(monitors, "grow, h pref:pref:300, top, wmax 100%")
-            add(synchronizers, "grow, h pref:pref:300, top, wmax 100%")
-            add(stacktrace, "push, grow, top, wmax 100%")
-        }
-
-        fun updateThreadInfo() {
-            isVisible = isViewable && isSelected
-
-            titleLabel.text = buildString {
-                tag("html") {
-                    tag("div") {
-                        tag("b", content = (thread?.state?.toString() ?: "NO THREAD"))
-                        append(" - ")
-                        append(percent.format(thread?.cpuUsage ?: 0.0))
-                    }
-                }
-            }
-
-            titleLabel.foreground = if (highlightCpu) {
-                threadHighlightColor
-            } else {
-                UIManager.getColor("Label.foreground")
-            }
-
-            blockerButton.blocker = thread?.blocker?.owner
-            detailsButton.isVisible = thread != null
-
-            monitors.apply {
-                isVisible = ShowEmptyValues.currentValue || thread?.lockedMonitors?.isNotEmpty() == true
-                if (isVisible) {
-                    itemCount = thread?.lockedMonitors?.size ?: 0
-                    text = thread?.lockedMonitors
-                        ?.joinToString(
-                            separator = "\n",
-                            prefix = "<html><pre>",
-                            postfix = "</pre></html>",
-                        ) { monitor ->
-                            if (monitor.frame == null) {
-                                "lock: ${monitor.lock}"
-                            } else {
-                                "lock: ${monitor.lock}\n${monitor.frame}"
-                            }
-                        }
-                }
-            }
-
-            synchronizers.apply {
-                isVisible = ShowEmptyValues.currentValue || thread?.lockedSynchronizers?.isNotEmpty() == true
-                if (isVisible) {
-                    itemCount = thread?.lockedSynchronizers?.size ?: 0
-                    text = thread?.lockedSynchronizers
-                        ?.joinToString(
-                            separator = "\n",
-                            prefix = "<html><pre>",
-                            postfix = "</pre></html>",
-                            transform = String::escapeHtml,
-                        )
-                }
-            }
-
-            stacktrace.apply {
-                isVisible = ShowEmptyValues.currentValue || thread?.stacktrace?.isNotEmpty() == true
-                if (isVisible) {
-                    itemCount = thread?.stacktrace?.size ?: 0
-                    text = thread?.stacktrace
-                        ?.joinToString(
-                            separator = "\n",
-                            prefix = "<html><pre>",
-                            postfix = "</pre></html>",
-                        ) { stackLine ->
-                            if (UseHyperlinks.currentValue) {
-                                stackLine.toBodyLine(version).let { (text, link) ->
-                                    if (link != null) {
-                                        """<a href="$link">$text</a>"""
-                                    } else {
-                                        text
-                                    }
-                                }
-                            } else {
-                                stackLine
-                            }
-                        }
-                    isHightlighted = highlightStacktrace
-                }
-            }
-
-            diffCheckBox.apply {
-                isSelected = false
-                isVisible = !thread?.stacktrace.isNullOrEmpty()
-            }
-        }
-    }
-
     companion object {
-        private val blockedIcon = FlatSVGIcon("icons/bx-block.svg").derive(12, 12)
-        private val detailIcon = FlatSVGIcon("icons/bx-link-external.svg").derive(12, 12)
-
-        private val percent = DecimalFormat("0.000'%'")
-
-        private val threadHighlightColor: Color
-            get() = UIManager.getColor("Component.warning.focusedBorderColor")
-
         private val nextIcon = FlatSVGIcon("icons/bx-chevron-right.svg")
         private val previousIcon = FlatSVGIcon("icons/bx-chevron-left.svg")
+
+        internal val threadHighlightColor: Color
+            get() = UIManager.getColor("Component.warning.focusedBorderColor")
     }
 }
