+++ src/main/kotlin/io/github/inductiveautomation/kindling/utils/Swing.kt
@@ -3,20 +3,14 @@ package io.github.inductiveautomation.kindling.utils
 import com.formdev.flatlaf.extras.FlatSVGIcon
 import com.github.weisj.jsvg.SVGDocument
 import com.github.weisj.jsvg.attributes.ViewBox
-import org.jdesktop.swingx.decorator.AbstractHighlighter
-import org.jdesktop.swingx.decorator.ColorHighlighter
-import org.jdesktop.swingx.decorator.ComponentAdapter
-import org.jdesktop.swingx.decorator.Highlighter
-import org.jdesktop.swingx.prompt.BuddySupport
 import java.awt.Color
-import kotlinx.coroutines.launch
 import java.awt.Component
 import java.awt.Container
 import java.awt.Point
 import java.awt.RenderingHints
 import java.awt.Toolkit
-import java.awt.event.KeyEvent
 import java.awt.datatransfer.Clipboard
+import java.awt.event.KeyEvent
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
 import java.awt.image.BufferedImage
@@ -27,7 +21,6 @@ import javax.swing.JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT
 import javax.swing.JFileChooser
 import javax.swing.JFrame
 import javax.swing.JPopupMenu
-import javax.swing.JTextArea
 import javax.swing.JScrollPane
 import javax.swing.JTextField
 import javax.swing.KeyStroke
@@ -39,7 +32,13 @@ import javax.swing.event.EventListenerList
 import javax.swing.text.Document
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.swing.Swing
+import org.jdesktop.swingx.decorator.AbstractHighlighter
+import org.jdesktop.swingx.decorator.ColorHighlighter
+import org.jdesktop.swingx.decorator.ComponentAdapter
+import org.jdesktop.swingx.decorator.Highlighter
+import org.jdesktop.swingx.prompt.BuddySupport
 
 /**
  * A common CoroutineScope bound to the event dispatch thread (see [Dispatchers.Swing]).
@@ -62,25 +61,6 @@ fun JFrame.dismissOnEscape() {
     )
     rootPane.getInputMap(WHEN_ANCESTOR_OF_FOCUSED_COMPONENT)
         .put(KeyStroke.getKeyStroke(KeyEvent.VK_ESCAPE, 0), "dismiss")
-}
-
-fun JTextArea.addLineHighlighter(
-    color: Color,
-    predicate: (line: String, lineNum: Int) -> Boolean,
-) {
-    if (text.isEmpty()) return
-    val highlighter = DefaultHighlighter.DefaultHighlightPainter(color)
-
-    for (lineNum in 0..<lineCount) {
-        val start = getLineStartOffset(lineNum)
-        val end = getLineEndOffset(lineNum)
-
-        val lineText = getText(start, end - start)
-
-        if (predicate(lineText, lineNum)) {
-            getHighlighter().addHighlight(start, end, highlighter)
-        }
-    }
 }
 
 inline fun <T : Component> T.attachPopupMenu(crossinline menuFn: T.(event: MouseEvent) -> JPopupMenu?) {
